fix(agent-runner): retry assigned sprite slots
